Add pryrc overrides (#2)
Previously appdev projects manually patched Pry behavior for the Rails console. This configuration was often forgotten about when new projects were created or when old projects were updated. Adding it as a setting to this gem will make projects easier to maintain.

This commit adds the ability to configure the Pry overrides without a .pryrc file.